chore(bindings): move tests into seperate dir
